Refactor project over singleton DB connection
- Move app logic from app.py to db_connection.py, with a class method for each function, eg. shorten_url, url_redirect, and a statistics section
- Add project config file to tidy up Flask parameters
- Refactor singleton DB connection in db_connector.py
