--- repository/db_conection.py
@@ -1,23 +1,0 @@
-class DBConnection(object):
-    connection = None
-
-    @classmethod
-    def get_connection(cls, new=False):
-        """Creates return new Singleton database connection"""
-        if new or not cls.connection:
-            cls.connection = DBConnector().create_connection()
-        return cls.connection
-
-    @classmethod
-    def execute_query(cls, query):
-        """execute query on singleton db connection"""
-        connection = cls.get_connection()
-        try:
-            cursor = connection.cursor()
-        except pyodbc.ProgrammingError:
-            connection = cls.get_connection(new=True)  # Create new connection
-            cursor = connection.cursor()
-        cursor.execute(query)
-        result = cursor.fetchall()
-        cursor.close()
-        return result
