+++ requirements.txt
@@ -1,2 +1,4 @@
-flask
-hashids
+flask~=2.2.3
+hashids~=1.3.1
+python-dotenv~=1.0.0
+pyodbc~=4.0.35
