+++ templates/index.html
@@ -6,7 +6,7 @@
 </head>
 <body>
 
-    <a href="{{ url_for('stats') }}">Statistics</a>
+    <a href="/">Statistics</a>
     <hr>
 
     <form action="" method="POST">
